feat: Add user creation script and authentication header service
- Implemented a standalone script for creating an admin user in the database using Axios for HTTP requests.
- Added an authentication header service to manage token retrieval for API requests.
- Created gamification services to manage sound effects, rewards, and task completion feedback.
- Developed objectives service for CRUD operations related to client objectives.
- Introduced profitability service for managing client profitability metrics and calculations.
- Added task impact service to identify and manage high-impact tasks based on the Pareto principle.
- Implemented timer service for tracking time spent on tasks and managing timers.
