+++ backend/package.json
@@ -16,16 +16,19 @@
     "@langchain/textsplitters": "^0.1.0",
     "bcrypt": "^5.1.0",
     "bcryptjs": "^2.4.3",
+    "bson": "^6.10.3",
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^4.18.2",
+    "express-mongo-sanitize": "^2.2.0",
     "glob": "^11.0.2",
+    "helmet": "^8.1.0",
     "jsonwebtoken": "^9.0.0",
     "mongodb": "^6.16.0",
     "mongoose": "^7.0.3",
     "multer": "^1.4.5-lts.2",
     "openai": "^4.96.0",
-    "bson": "^6.10.3"
+    "winston": "^3.17.0"
   },
   "devDependencies": {
     "jest": "^29.5.0",
