+++ backend/tests/test-mongo-image.js
