+++ frontend/src/index.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { store, resetStore } from './store';
+import { store, resetStore } from './store/index';
 import './index.css';
 
 // Composant ErrorBoundary pour capturer les erreurs
