+++ frontend/src/services/gamificationServices.ts.backup
